anchor/requirements.txt
django
django-polaris
gunicorn
django-cors-headers
python-dateutil
stellar-sdk

